conversion fix float instead of int

+++ neurons/miners/bitcoin/funds_flow/indexer.py
@@ -86,7 +86,7 @@ def index_blocks(_bitcoin_node, _graph_creator, _graph_indexer):
                 # indexer flooding prevention
                 threshold = int(os.getenv('BLOCK_PROCESSING_TRANSACTION_THRESHOLD', 10))
                 if num_transactions < threshold:
-                    delay = int(os.getenv('BLOCK_PROCESSING_DELAY', .1))
+                    delay = float(os.getenv('BLOCK_PROCESSING_DELAY', .1))
                     logger.info(f"Block tx count below {threshold}, slowing down indexing by {delay} seconds to prevent flooding.")
                     time.sleep(delay)
